requirements.txt
APScheduler==3.0.5
boto==2.38.0
ddt==1.0.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-bootstrap-form==3.2
django-braces==1.8.1
django-filter==0.10.0
django-model-utils==2.4
django-registration==2.0.3
django-sslify==0.2.7
django-storages==1.4
django-toolbelt==0.0.1
djangorestframework==3.1.3
futures==3.0.3
gunicorn==19.3.0
jsonfield==1.0.3
oauthlib==1.0.3
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
static3==0.6.1
stripe==1.29.1
tzlocal==1.2
wheel==0.24.0
whitenoise==2.0.2
